requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0
mysql-connector-python==9.6.0
gunicorn==25.1.0
